+++ src/world/terraformation/landscaping.rs
@@ -2,10 +2,7 @@ use bevy::{prelude::*, utils::HashSet};
 
 use projekto_core::{chunk, query};
 
-use super::{
-    prelude::{BatchChunkCmdRes, WorldRes},
-    TerraformationCenter, TerraformationConfig,
-};
+use super::{prelude::BatchChunkCmdRes, TerraformationCenter, TerraformationConfig, genesis::ChunkKindRes};
 
 pub(super) struct LandscapingPlugin;
 
@@ -23,16 +20,12 @@ struct UpdateLandscapeMeta {
 fn update_landscape(
     time: Res<Time>,
     config: Res<TerraformationConfig>,
-    world_res: Res<WorldRes>,
+    kinds: Res<ChunkKindRes>,
     mut meta: Local<UpdateLandscapeMeta>,
     mut batch: ResMut<BatchChunkCmdRes>,
     q: Query<&Transform, With<TerraformationCenter>>,
 ) {
     let mut _perf = perf_fn!();
-
-    if !world_res.is_ready() {
-        return;
-    }
 
     let center = match q.get_single() {
         Ok(t) => chunk::to_local(t.translation),
@@ -56,7 +49,7 @@ fn update_landscape(
         let end = center + radius;
 
         let visible_range = query::range_inclusive(begin, end).collect::<HashSet<_>>();
-        let existing_chunks = HashSet::from_iter(world_res.list_chunks().into_iter());
+        let existing_chunks = HashSet::from_iter(kinds.list_chunks().into_iter());
 
         visible_range
             .iter()
